new branch with additions of the shuffle function made to the deck class
