+++ WarController.java
@@ -2,7 +2,7 @@
 
 public class WarController {
     WarController() {
-        Hand player1 = new Hand();
+        Deck player1 = new Deck();
         
 
     }
